Implement image padding role
We're having issues where Imager's comparison is grumpy about images
that don't have identical sizes. In particular, because Firefox takes an
image of the entire website (and not just the current viewport), AND
because content changes from page load to page load, it's possible for
the images to different by a pixel or two each time. This means the
comparison blows up.

This role gives us a subroutine that takes two images and gives them
back to us padded to be the same size, so we no longer get the 'Dying:
Images must be the same size' explosions.