Rework target attr to be processed in BUILDARGS
The target logic really needs to be applied as soon as
possible. Otherwise, we get into really confusing states where we're not
sure whether we already applied the target cropping, or what we should
be applying the exclusion rectangles to.

In fact, this will probably break the interaction between exclusions and
targetting. On the other hand, exclusions and targetting never worked
together at all, so we're not really losing ground. Exclusion on its own
will still work, and now targetting works as expected as well.